+++ contracts/TodoListSolidity.sol
@@ -21,21 +21,29 @@ contract List {
         owner = msg.sender;
     }
 
-    function addTodo(string calldata _title, string calldata _description) public onlyOwner {
+    function addTodo(
+        string calldata _title,
+        string calldata _description
+    ) public onlyOwner {
         todos.push(
             Todo({title: _title, description: _description, completed: false})
         );
     }
 
-    function changeTodoTitle(string calldata _newTitle, uint index) public onlyOwner {
+    function changeTodoTitle(
+        string calldata _newTitle,
+        uint256 index
+    ) public onlyOwner {
         todos[index].title = _newTitle;
     }
 
-    function changeTodoStatus(uint index) public onlyOwner {
+    function changeTodoStatus(uint256 index) public onlyOwner {
         todos[index].completed = !todos[index].completed;
     }
 
-    function getTodo(uint index) public view onlyOwner returns (string memory, string memory, bool) {
+    function getTodo(
+        uint256 index
+    ) public view onlyOwner returns (string memory, string memory, bool) {
         Todo storage myTodo = todos[index];
 
         return (myTodo.title, myTodo.description, myTodo.completed);
